rename: change project name from aerospike-ui to aerospike-py-admin-ui
- Rename Python package directory: aerospike_ui_api -> aerospike_py_admin_ui_api
- Update all Python imports across 19 source files
- Update pyproject.toml, uv.lock, Dockerfile, docker-compose.yml
- Update frontend package.json, package-lock.json, localStorage key
- Update README.md, CLAUDE.md documentation
- Fix CD build: add README.md to Dockerfile COPY step
