+++ backend/README.md
@@ -1,3 +1,3 @@
-# Aerospike UI Backend API
+# Aerospike Py Admin UI Backend API
 
-FastAPI backend for Aerospike UI.
+FastAPI backend for Aerospike Py Admin UI.
