+++ backend/src/aerospike_py_admin_ui_api/mock_data/udfs.py
@@ -1,4 +1,4 @@
-from aerospike_ui_api.models.udf import UDFModule
+from aerospike_py_admin_ui_api.models.udf import UDFModule
 
 RECORD_UDF_SOURCE = """\
 -- record_udf.lua
