+++ backend/src/aerospike_py_admin_ui_api/models/record.py
