+++ backend/src/aerospike_py_admin_ui_api/routers/indexes.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from fastapi import APIRouter, HTTPException
 
-from aerospike_ui_api import store
-from aerospike_ui_api.models.index import CreateIndexRequest, SecondaryIndex
+from aerospike_py_admin_ui_api import store
+from aerospike_py_admin_ui_api.models.index import CreateIndexRequest, SecondaryIndex
 
 router = APIRouter(prefix="/api/indexes", tags=["indexes"])
